Set C binding for unix_getcwd_job in its own C file
