+++ src/unix/jbuild
@@ -204,6 +204,7 @@
     unix_getaddrinfo_job
     unix_getnameinfo_job
     unix_bind_job
+    unix_getcwd_job
     not_available
   ))
   (install_c_headers (
