+++ src/unix/windows_c/not_available.c
@@ -91,4 +91,5 @@ LWT_NOT_AVAILABLE2(unix_getservbyport_job)
 LWT_NOT_AVAILABLE3(unix_getaddrinfo_job)
 LWT_NOT_AVAILABLE2(unix_getnameinfo_job)
 LWT_NOT_AVAILABLE2(unix_bind_job)
+LWT_NOT_AVAILABLE1(unix_getcwd_job)
 #endif
